# docs/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "My Activity Stats API",
    "version": "1.4.0",
    "description": "Canonical cross-backend HTTP contract. Existing URLs are preserved. GET endpoints also support HEAD (same status and headers, no body). Unsupported methods return 405 with Allow; unknown API URLs return 404 JSON. Every API response includes X-Request-Id. Object schemas with additionalProperties remain intentionally partial until domain DTO coverage is extended."
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "paths": {
...
    "/api/activities": {
      "get": {
        "operationId": "listActivities",
        "tags": [
          "activities"
        ],
        "responses": {
          "200": {
            "description": "Activities",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ActivitySummary"
                  }
                }
              }
            },
            "headers": {
              "X-Request-Id": {
                "$ref": "#/components/headers/RequestId"
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "405": {
            "$ref": "#/components/responses/MethodNotAllowed"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        },
        "parameters": [
          {
            "name": "activityType",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Strava activity types separated by underscores, for example Ride_GravelRide. Case-sensitive."
          },
          {
            "name": "year",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer"
            },
            "description": "Calendar year; omitted means all years except on charts, where it is required."
          }
        ]
      },
      "parameters": [
        {
          "$ref": "#/components/parameters/RequestId"
        }
      ]
    },
    "/api/activities/csv": {
      "get": {
        "operationId": "exportActivitiesCsv",
        "tags": [
          "activities"
        ],
        "responses": {
          "200": {
            "description": "CSV export",
            "content": {
              "text/csv": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "headers": {
              "X-Request-Id": {
                "$ref": "#/components/headers/RequestId"
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "405": {
            "$ref": "#/components/responses/MethodNotAllowed"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        },
        "parameters": [
          {
            "name": "activityType",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Strava activity types separated by underscores, for example Ride_GravelRide. Case-sensitive."
          },
          {
            "name": "year",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer"
            },
            "description": "Calendar year; omitted means all years except on charts, where it is required."
          }
        ]
      },
      "parameters": [
        {
          "$ref": "#/components/parameters/RequestId"
        }
      ]
    },
    "/api/activities/{activityId}": {
      "parameters": [
        {
          "$ref": "#/components/parameters/ActivityId"
        },
        {
          "$ref": "#/components/parameters/RequestId"
        }
      ],
      "get": {
        "operationId": "getActivity",
        "tags": [
          "activities"
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/JsonSuccess"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "405": {
            "$ref": "#/components/responses/MethodNotAllowed"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        },
        "parameters": [
          {
            "name": "version",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "raw returns uncorrected data; any other value uses corrected data."
          }
        ]
      }
    },
...
    "/api/maps/gpx": {
      "get": {
        "operationId": "getMapTracks",
        "tags": [
          "maps"
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/JsonArraySuccess"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "405": {
            "$ref": "#/components/responses/MethodNotAllowed"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        },
        "parameters": [
          {
            "name": "activityType",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Strava activity types separated by underscores, for example Ride_GravelRide. Case-sensitive."
          },
          {
            "name": "year",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer"
            },
            "description": "Calendar year; omitted means all years except on charts, where it is required."
          }
        ]
      },
      "parameters": [
        {
          "$ref": "#/components/parameters/RequestId"
        }
      ]
    },
    "/api/maps/passages": {
      "get": {
        "operationId": "getMapPassages",
        "tags": [
          "maps"
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/JsonSuccess"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "405": {
            "$ref": "#/components/responses/MethodNotAllowed"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        },
        "parameters": [
          {
            "name": "activityType",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Strava activity types separated by underscores, for example Ride_GravelRide. Case-sensitive."
          },
          {
            "name": "year",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer"
            },
            "description": "Calendar year; omitted means all years except on charts, where it is required."
          }
        ]
      },
      "parameters": [
        {
          "$ref": "#/components/parameters/RequestId"
        }
      ]
    },
...
  },
  "components": {
    "parameters": {
      "ActivityId": {
        "name": "activityId",
        "in": "path",
        "required": true,
        "schema": {
          "type": "integer",
          "format": "int64"
        }
      },
...
      "RequestId": {
        "name": "X-Request-Id",
        "in": "header",
        "required": false,
        "description": "Echoed when it matches the pattern; otherwise a new ID is generated.",
        "schema": {
          "type": "string",
          "pattern": "^[A-Za-z0-9._:-]{1,128}$"
        }
      }
    },
    "responses": {
      "JsonSuccess": {
        "description": "Successful JSON response",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          }
        },
        "headers": {
          "X-Request-Id": {
            "$ref": "#/components/headers/RequestId"
          }
        }
      },
...
      "JsonArraySuccess": {
        "description": "JSON collection (empty collection is []).",
        "content": {
          "application/json": {
            "schema": {
              "type": "array",
              "items": {
                "type": "object",
                "additionalProperties": true
              }
            }
          }
        },
        "headers": {
          "X-Request-Id": {
            "$ref": "#/components/headers/RequestId"
          }
        }
      },
      "BadRequest": {
        "description": "Invalid parameters or malformed JSON",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            }
          }
        },
        "headers": {
          "X-Request-Id": {
            "$ref": "#/components/headers/RequestId"
          }
        }
      },
      "NotFound": {
        "description": "Resource or API endpoint not found",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            }
          }
        },
        "headers": {
          "X-Request-Id": {
            "$ref": "#/components/headers/RequestId"
          }
        }
      },
      "MethodNotAllowed": {
        "description": "HTTP method not supported",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            }
          }
        },
        "headers": {
          "Allow": {
            "description": "Supported HTTP methods",
            "schema": {
              "type": "string"
            }
          },
          "X-Request-Id": {
            "$ref": "#/components/headers/RequestId"
          }
        }
      },
      "InternalError": {
        "description": "Internal server error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            }
          }
        },
        "headers": {
          "X-Request-Id": {
            "$ref": "#/components/headers/RequestId"
          }
        }
      },
...
    },
    "schemas": {
      "ApiError": {
        "type": "object",
        "required": [
          "message",
          "description",
          "code",
          "requestId"
        ],
        "properties": {
          "message": {
            "type": "string"
          },
          "code": {
            "type": "integer",
            "description": "Legacy application code: 1 for API errors, 403 for an origin rejection. Use the HTTP status for error classification."
          },
          "description": {
            "type": "string"
          },
          "requestId": {
            "type": "string"
          }
        }
      },
...
      "ActivitySummary": {
        "type": "object",
        "required": [
          "id",
          "name",
          "type"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "date": {
            "type": "string",
            "nullable": true
          },
          "distance": {
            "type": "number",
            "nullable": true
          }
        }
      },
...
    },
    "headers": {
      "RequestId": {
        "description": "Request correlation identifier, also included in JSON errors.",
        "schema": {
          "type": "string"
        }
      }
    }
  }
}